LayoutTests/fast/dom/SelectorAPI/caseTagX-expected.txt

PASS document.querySelector('div SPAN') is null
PASS successfullyParsed is true

TEST COMPLETE
